Sleep Stats up and running. This may move to just an overall stats screen and then each data will have only a chart?

+++ app/src/main/java/app/collegeorganizer/data/ComparisonOperator_Ratio.java
@@ -1,10 +1,10 @@
 package app.collegeorganizer.data;
 
 public class ComparisonOperator_Ratio {
-    private int data1;
-    private int data2;
+    private float data1;
+    private float data2;
 
-    public ComparisonOperator_Ratio(int data1, int data2) {
+    public ComparisonOperator_Ratio(float data1, float data2) {
         this.data1 = data1;
         this.data2 = data2;
     }
